backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
boto3==1.34.0
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
numpy==1.24.3
ffmpeg-python==0.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
stripe==7.8.0
sendgrid==6.10.0
python-dotenv==1.0.0
